Module sui_indexer::db

Re-exports

pub use setup_postgres::reset_database;

pub use setup_postgres::run_migrations;

Modules

__diesel_schema_migrations

setup_postgres

Structs

ConnectionConfig

ConnectionPoolConfig

Functions

check_db_migration_consistency

Checks that the local migration scripts is a prefix of the records in the database. This allows us run migration scripts against a DB at anytime, without worrying about existing readers fail over. We do however need to make sure that whenever we are deploying a new version of either reader or writer, we must first run migration scripts to ensure that there is not more local scripts than in the DB record.

check_prunable_tables_valid

Check that prunable tables exist in the database.