sui_indexer::schema::event_emit_module
Module dsl

Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Re-exports

pub use super::columns::package;

pub use super::columns::module;

pub use super::columns::tx_sequence_number;

pub use super::columns::event_sequence_number;

pub use super::columns::sender;

pub use super::table as event_emit_module;