Module sui_indexer::schema::objects_history_partition_0

Modules

columns

Contains all of the columns of this table

dsl

Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Structs

checkpoint_sequence_number

coin_balance

coin_type

df_kind

df_name

df_object_id

df_object_type

object_digest

object_id

object_status

object_type

object_type_module

object_type_name

object_type_package

object_version

owner_id

owner_type

serialized_object

star

Represents table_name.*, which is sometimes needed for efficient count queries. It cannot be used in place of all_columns, and has a SqlType of () to prevent it being used that way

table

The actual table struct

Constants

all_columns

A tuple of all of the columns on this table

Type Aliases

BoxedQuery

Helper type for representing a boxed query from this table

SqlType

The SQL type of all of the columns on this table