Module sui_indexer::handlers

Modules

checkpoint_handler

committer

objects_snapshot_handler

pruner

tx_processor

Structs

CheckpointDataToCommit

CommitterTablesIter

An iterator over the variants of Self

CommitterWatermark

The indexer writer operates on checkpoint data, which contains information on the current epoch, checkpoint, and transaction. These three numbers form the watermark upper bound for each committed table. The reader and pruner are responsible for determining which of the three units will be used for a particular table.

CommonHandler

EpochToCommit

ObjectsSnapshotHandlerTablesIter

An iterator over the variants of Self

TransactionObjectChangesToCommit

Enums

CommitterTables

Enum representing tables that the committer handler writes to.

ObjectsSnapshotHandlerTables

Enum representing tables that the objects snapshot handler writes to.

Traits

Handler