sui_indexer
Module test_utils

Structs

SuiTransactionBlockResponseBuilder

Functions

set_up

Set up a test indexer fetching from a REST endpoint served by the given Simulacrum.

set_up_with_start_and_end_checkpoints

start_indexer_jsonrpc_for_testing

Wrapper over Indexer::start_reader to make it easier to configure an indexer jsonrpc reader for testing.

start_indexer_writer_for_testing

Wrapper over Indexer::start_writer_with_config to make it easier to configure an indexer writer for testing. If the config options are null, default values that have historically worked for testing will be used.

wait_for_checkpoint

Wait for the indexer to catch up to the given checkpoint sequence number.

wait_for_objects_snapshot

Wait for the indexer to catch up to the given checkpoint sequence number for objects snapshot.